Improving the Data and the R Script
Although the R script is still based on the original 21 files, there
are now 47 debates available.  Major updates include changes to the
corrections object in the i2-getstarted.py file as well as changes to
the script 20140409 R..R which include analysis of the polling results
as well as preliminary correlations between outcome and LIWC
characteristics.